Split `script.py` file
