+++ pymodd/script.py
@@ -1,209 +1,14 @@
-import os
 import ast
 import inspect
-import json
-import random
-import string
 import textwrap
-from _ast import AnnAssign, Assign, Delete
 from enum import Enum
+from _ast import AnnAssign, Assign, Delete
 
-from caseconverter import camelcase, snakecase
+from caseconverter import snakecase
 
 import pymodd
 from pymodd import _pymodd_helper
-
-
-class Base():
-    def to_dict(self):
-        raise NotImplementedError('to_dict method not implemented')
-
-
-class Game(Base):
-    def __init__(self, json_file_path, game_variable_classes, project_globals_data):
-        self.project_globals_data = project_globals_data
-        with open(json_file_path, 'r') as file:
-            data = json.load(file)
-        self.name = data.get('title')
-        self.data = data
-        self.entity_scripts = []
-        self.scripts = []
-        self._update_data_with_variable_classes(game_variable_classes)
-        self._build()
-        # set position of scripts inside game
-        for i, script in enumerate(self.scripts):
-            script.set_position(i, None)
-
-    def _update_data_with_variable_classes(self, variable_classes):
-        # pull variable objects out from each class and place them in categories
-        variable_category_to_variables = {}
-        for klass in variable_classes:
-            variable_category = variable_category_name_from_variable_class_name(
-                klass.__name__)
-            variable_category_to_variables.setdefault(variable_category, [])
-            class_vars = [item for item in vars(klass)
-                          if not item.startswith('_') and not callable(item)]
-            for var in class_vars:
-                variable_category_to_variables[variable_category].append(getattr(
-                    klass, var))
-
-        # update game data with modified variable objects
-        for variable_category, variables in variable_category_to_variables.items():
-            if variable_category not in self.data['data'].keys():
-                continue
-            category_data = self.data['data'][variable_category]
-            unincluded_variable_ids = list(
-                category_data.keys())
-            for variable in variables:
-                category_contains_variable = variable.id in category_data.keys()
-                category_data[variable.id] = variable.updated_data_with_user_provided_values(
-                    category_data[variable.id] if category_contains_variable else variable.get_template_data())
-                if variable.id in unincluded_variable_ids:
-                    unincluded_variable_ids.remove(variable.id)
-            # remove variables no longer included
-            for unincluded_variable_id in unincluded_variable_ids:
-                category_data.pop(unincluded_variable_id)
-
-    def _build():
-        pass
-
-    def to_dict(self):
-        # update global scripts
-        self.data['data']['scripts'] = self.flatten_scripts_data(self.scripts)
-
-        # update data of each entity_type
-        for entity_script in self.entity_scripts:
-            entity_category, entity_id = f'{camelcase(entity_script.entity_type.__class__.__name__)}s', entity_script.entity_type.id
-            entity_data = self.data['data'][entity_category][entity_id]
-
-            # update entity scripts
-            entity_data['scripts'] = self.flatten_scripts_data(
-                entity_script.scripts)
-
-            # update entity keybindings
-            entity_keybindings_data = entity_data['controls']['abilities']
-            unincluded_keys = list(entity_keybindings_data.keys())
-            for (key, scripts) in entity_script.keybindings.items():
-                entity_keybindings_data[key.value] = scripts.to_dict(
-                    entity_keybindings_data.get(key.value))
-                if key.value in unincluded_keys:
-                    unincluded_keys.remove(key.value)
-            # remove keybindings no longer included
-            for unincluded_key in unincluded_keys:
-                if unincluded_key in ['lookWheel', 'movementWheel']:
-                    continue
-                entity_keybindings_data.pop(unincluded_key)
-        return self.data
-
-    def flatten_scripts_data(self, scripts):
-        '''Takes all scripts out of folders, transforms them into json, and returns one dictionary with all of the game's script
-
-        Returns:
-            dict(str, dict): keys are script keys, values are datas of scripts
-        '''
-        flattened_scripts = {}
-        scripts_queue = scripts.copy()
-        while len(scripts_queue) > 0:
-            script = scripts_queue.pop(0)
-            # add folder's scripts to the queue
-            if isinstance(script, Folder):
-                scripts_queue += script.scripts
-            script_data = script.to_dict(self.project_globals_data)
-            flattened_scripts[script_data['key']] = script_data
-        return flattened_scripts
-
-
-def variable_category_name_from_variable_class_name(variable_class_name):
-    if variable_class_name == 'EntityVariables':
-        return 'entityTypeVariables'
-    elif variable_class_name == 'PlayerVariables':
-        return 'playerTypeVariables'
-    elif variable_class_name == 'Musics':
-        return 'music'
-    elif variable_class_name in ['Regions', 'ItemTypeGroups', 'UnitTypeGroups']:
-        return 'variables'
-    else:
-        return camelcase(variable_class_name)
-
-
-class EntityScripts(Game):
-    def __init__(self):
-        self.entity_type = None
-        self.keybindings = {}  # dict of Keys to KeyBehavior
-        self.scripts = []
-        self._build()
-        # set position of scripts inside entity_scripts
-        for i, script in enumerate(self.scripts):
-            script.set_position(i, None)
-
-
-class KeyBehavior(Base):
-    def __init__(self, key_down_script=None, key_up_script=None):
-        self.key_down_script_key, self.is_key_down_script_entity_script = "", False
-        self.key_up_script_key, self.is_key_up_script_entity_script = "", False
-        if key_down_script is not None:
-            self.key_down_script_key = key_down_script.key
-            self.is_key_down_script_entity_script = os.path.split(
-                key_down_script.build_func_source_file)[-1] == 'entity_scripts.py'
-        if key_up_script is not None:
-            self.key_up_script_key = key_up_script.key
-            self.is_key_up_script_entity_script = os.path.split(
-                key_up_script.build_func_source_file)[-1] == 'entity_scripts.py'
-
-    def to_dict(self, old_data: None):
-        # update the old data with new data if it is provided
-        data = old_data if old_data is not None else self.get_template_data()
-        data['keyDown']['scriptName'] = self.key_down_script_key
-        data['keyDown']['isEntityScript'] = self.is_key_down_script_entity_script
-        data['keyUp']['scriptName'] = self.key_up_script_key
-        data['keyUp']['isEntityScript'] = self.is_key_up_script_entity_script
-        return data
-
-    def get_template_data(self):
-        return {
-            "keyDown": {
-                "scriptName": "",
-                "isEntityScript": False,
-                "cost": {},
-            },
-            "keyUp": {
-                "scriptName": "",
-                "isEntityScript": False,
-                "cost": {},
-            }
-        }
-
-
-class File(Base):
-    def __init__(self):
-        self.name = None
-        self.key = None
-        self.parent = None
-        self.order = 0
-
-    def set_position(self, order, parent):
-        self.order = order
-        self.parent = parent
-
-
-class Folder(File):
-    def __init__(self, name, scripts: list):
-        super().__init__()
-        self.name = name
-        self.key = generate_random_key()
-        self.scripts = scripts
-        # set position of scripts inside the folder
-        for i, script in enumerate(scripts):
-            script.set_position(i, self.key)
-
-    def to_dict(self):
-        return {
-            'key': self.key,
-            'folderName': self.name,
-            'parent': self.parent,
-            'order': self.order,
-            'expanded': True
-        }
+from pymodd.game import Base, File, generate_random_key
 
 
 class Script(File):
@@ -468,10 +273,6 @@ class ScriptActionsCompiler(ast.NodeVisitor):
         return eval(code, self.project_globals_data, self.get_current_locals_data())
 
 
-def generate_random_key():
-    return ''.join(random.choices(string.ascii_letters + string.digits, k=10))
-
-
 def to_dict(obj):
     if isinstance(obj, Base):
         return obj.to_dict()
@@ -564,41 +365,3 @@ class Flip(Enum):
     HORIZONTAL = 'horizontal'
     VERTICAL = 'vertical'
     BOTH = 'both'
-
-
-class Key(Enum):
-    LEFT_CLICK = 'button1'
-    RIGHT_CLICK = 'button3'
-    SPACE = 'space'
-    ENTER = 'enter'
-    ESCAPE = 'escape'
-    UP = 'up'
-    DOWN = 'down'
-    RIGHT = 'right'
-    LEFT = 'left'
-    A = 'a'
-    B = 'b'
-    C = 'c'
-    D = 'd'
-    E = 'e'
-    F = 'f'
-    G = 'g'
-    H = 'h'
-    I = 'i'
-    J = 'j'
-    K = 'k'
-    L = 'l'
-    M = 'm'
-    N = 'n'
-    O = 'o'
-    P = 'p'
-    Q = 'q'
-    R = 'r'
-    S = 's'
-    T = 't'
-    U = 'u'
-    V = 'v'
-    W = 'w'
-    X = 'x'
-    Y = 'y'
-    Z = 'z'
